second

+++ src/services/User.ts
@@ -2,6 +2,10 @@ import User from "../models/User";
 
 export const createUser = async (data: any) => {
   try {
+    const user = await User.findOne({email:data.email});
+    if (user) {
+      throw new Error("User already exists");
+    }
     return await User.create(data);
   } catch (error:any) {
     throw new Error(error);
@@ -17,6 +21,10 @@ export const getAllUsers = async () => {
     };
     export const getUser = async (id: string) => {
         try {
+            const user = await User.findById(id);
+            if (!user) {
+                throw new Error("User not found");
+            }
             return await User.findById(id);
         } catch (error:any) {
             throw new Error(error);
@@ -24,6 +32,10 @@ export const getAllUsers = async () => {
     };
     export const updateUser = async (id: string, data: any) => {
         try {
+            const user = await User.findById(id);
+            if (!user) {
+                throw new Error("User not found");
+            }
             return await User.findByIdAndUpdate(id, data, { new: true });
         }
         catch (error:any) {
@@ -32,6 +44,10 @@ export const getAllUsers = async () => {
     };
     export const deleteUser = async (id: string) => {
         try {
+            const user = await User.findById(id);
+            if (!user) {
+                throw new Error("User not found");
+            }
             return await User.findByIdAndDelete(id);
         }
         catch (error:any) {
